requirements.txt
tensorflow>=2.11
numpy
matplotlib
seaborn
scikit-learn
pillow
tk